backend/ai-kb-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.9.1
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.7
python-multipart==0.0.20
redis==6.0.0
qdrant-client==1.14.2
PyMuPDF==1.26.0
python-docx==1.1.2
